Add files via upload
here are 2 variations on a quick-sort code that using recursion for one and a loop and stack for the other. the idea here is to partial sort from a multidimensional array/vector. Its fully operational and there is a noticeable difference in cpu/memory usage when ran with the performance profiler in visual studios. I have one ran them with 10 thousand items to sort so far.